requirements_all_integrations.txt
# Complete Integration Dependencies for Agentic Workflow Platform
# Install with: pip install -r requirements_all_integrations.txt

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ============================================================================
# AWS & CLOUD
# ============================================================================
boto3==1.34.34
botocore==1.34.34

# ============================================================================
# TASK QUEUE & CACHING
# ============================================================================
redis==5.0.1
celery==5.3.4

# ============================================================================
# LLM PROVIDERS
# ============================================================================
anthropic==0.18.1
openai==1.12.0
langchain==0.1.10
langchain-anthropic==0.1.6
langchain-openai==0.0.8
langchain-community==0.0.24

# ============================================================================
# MULTI-AGENT FRAMEWORKS
# ============================================================================
crewai==0.11.0
crewai-tools==0.2.6
langgraph==0.0.26

# ============================================================================
# MEMORY SYSTEMS
# ============================================================================
mem0ai==0.0.8
zep-python==2.0.0

# ============================================================================
# VECTOR DATABASES & EMBEDDINGS
# ============================================================================
pinecone-client==3.1.0
weaviate-client==4.4.4

# ============================================================================
# WEB RESEARCH & SCRAPING
# ============================================================================
requests==2.31.0
beautifulsoup4==4.12.3
# Tavily, Firecrawl, Serper use their own APIs via requests

# ============================================================================
# DOCUMENT PROCESSING
# ============================================================================
pypdf2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
pillow==10.2.0
pdfplumber==0.10.4

# ============================================================================
# CODE EXECUTION
# ============================================================================
e2b-code-interpreter==0.0.7

# ============================================================================
# LLM ROUTING & OBSERVABILITY
# ============================================================================
litellm==1.30.0
langsmith==0.1.7
arize-phoenix==5.12.0  # Updated to latest stable version

# ============================================================================
# COMMUNICATION INTEGRATIONS
# ============================================================================
slack-sdk==3.27.0
twilio==9.0.0
discord.py==2.3.2
# MS Teams uses webhooks (requests library)

# ============================================================================
# EMAIL
# ============================================================================
google-api-python-client==2.119.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ============================================================================
# PAYMENT PROCESSING
# ============================================================================
stripe==8.7.0
# PayPal SDK can be added if needed

# ============================================================================
# PROJECT MANAGEMENT
# ============================================================================
PyGithub==2.2.0
jira==3.6.0
# Linear API uses requests

# ============================================================================
# CRM & MARKETING
# ============================================================================
hubspot-api-client==9.0.0
# Google Analytics (ga4) API
google-analytics-data==0.18.5

# ============================================================================
# CALENDAR & SCHEDULING
# ============================================================================
google-api-python-client==2.119.0  # Already listed above
APScheduler==3.10.4

# ============================================================================
# TESTING & DEVELOPMENT
# ============================================================================
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
faker==24.0.0

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.1
pyyaml==6.0.1
aiohttp==3.9.3
websockets==12.0

# ============================================================================
# OPTIONAL: AWS BEDROCK (if using AWS-hosted models)
# ============================================================================
# langchain-aws==0.0.1

# ============================================================================
# OPTIONAL: ADDITIONAL MODELS
# ============================================================================
# cohere==4.47
# replicate==0.24.0
# together==0.2.8

# ============================================================================
# OPTIONAL: OBSERVABILITY & MONITORING
# ============================================================================
# prometheus-client==0.20.0
# sentry-sdk==1.40.6

# ============================================================================
# NOTES
# ============================================================================
# 1. Some services (Firecrawl, Tavily, Serper) don't have Python SDKs
#    and use direct API calls via requests
#
# 2. Mock clients are included in integration files - no API keys needed
#    for development and testing
#
# 3. Optional dependencies commented out - uncomment as needed
#
# 4. Version pinning ensures compatibility - update carefully
#
# 5. Install in virtual environment:
#    python -m venv venv
#    source venv/bin/activate  # or venv\Scripts\activate on Windows
#    pip install -r requirements_all_integrations.txt
